+++ test/context_menu_android_test.dart
@@ -1,27 +1,41 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
-import 'package:context_menu_android/context_menu_android.dart';
+import 'package:context_menu_android/ios_style_context_menu.dart';
 
 void main() {
-  testWidgets('Context menu renders child correctly', (
-    WidgetTester tester,
-  ) async {
+  testWidgets('Context menu renders child correctly', (WidgetTester tester) async {
     final testKey = const Key('my_test_widget');
 
     await tester.pumpWidget(
       MaterialApp(
         home: Builder(
           builder: (context) {
-            return iOSStyleContextMenu(
-              context,
-              child: Container(key: testKey),
-              actions: [],
+            return Scaffold(
+              body: Center(
+                child: ElevatedButton(
+                  onPressed: () {
+                    showDialog<Widget>(
+                      context: context,
+                      builder: (_) => IosStyleContextMenu(
+                        actions: const [],
+                        child: Container(key: testKey),
+                      ),
+                    );
+                  },
+                  child: const Text('Open Menu'),
+                ),
+              ),
             );
           },
         ),
       ),
     );
 
+    // اضغط على الزر لعرض القائمة
+    await tester.tap(find.text('Open Menu'));
+    await tester.pumpAndSettle();
+
+    // تحقق من وجود الودجت داخل القائمة
     expect(find.byKey(testKey), findsOneWidget);
   });
 }